+++ src/main/java/com/skplanet/storeplatform/sac/display/vod/vo/VodDetail.java
@@ -146,9 +146,9 @@ public class VodDetail extends CommonInfo {
 	// Accrual
 	// ---------------------------------------------------
 	/** 참여자수 */
-	private Double paticpersCnt;
+	private Integer paticpersCnt;
 	/** 구매수 */
-	private Double prchsCnt;
+	private Integer prchsCnt;
 	/** 평균 평가 점수 */
 	private Double avgEvluScore;
 
@@ -460,7 +460,7 @@ public class VodDetail extends CommonInfo {
 	 *
 	 * @return 참여자수
 	 */
-	public Double getPaticpersCnt() {
+	public Integer getPaticpersCnt() {
 		return this.paticpersCnt;
 	}
 
@@ -470,7 +470,7 @@ public class VodDetail extends CommonInfo {
 	 * @param paticpersCnt
 	 *            참여자수
 	 */
-	public void setPaticpersCnt(Double paticpersCnt) {
+	public void setPaticpersCnt(Integer paticpersCnt) {
 		this.paticpersCnt = paticpersCnt;
 	}
 
@@ -479,7 +479,7 @@ public class VodDetail extends CommonInfo {
 	 *
 	 * @return 구매수
 	 */
-	public Double getPrchsCnt() {
+	public Integer getPrchsCnt() {
 		return this.prchsCnt;
 	}
 
@@ -489,7 +489,7 @@ public class VodDetail extends CommonInfo {
 	 * @param prchsCnt
 	 *            구매수
 	 */
-	public void setPrchsCnt(Double prchsCnt) {
+	public void setPrchsCnt(Integer prchsCnt) {
 		this.prchsCnt = prchsCnt;
 	}
